requirements.txt
django==3.2.4
PyYAML==5.4.1
#rest framework
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
django-oauth-toolkit==1.5.0

psycopg2-binary==2.8.6
django-countries==7.2.1
numpy==1.20.3
Pillow==8.2.0
django-language-field==0.0.3
django-computed-property==0.3.0
tqdm==4.61.0
fuzzysearch==0.7.3

django-prometheus==2.1.0

drf_spectacular==0.17.2
